+++ apps/demo-app/src/views/HomeView.vue
@@ -1,5 +1,6 @@
 <script lang="ts" setup>
 import emojiHappy2 from "@sit-onyx/icons/emoji-happy-2.svg?raw";
+import { useStorage } from "@vueuse/core";
 import {
   DENSITIES,
   OnyxAvatar,
@@ -24,6 +25,7 @@ import {
   OnyxTimer,
   OnyxTooltip,
   normalizedIncludes,
+  useToast,
   type SelectOption,
 } from "sit-onyx";
 import { capitalize, computed, ref } from "vue";
@@ -52,6 +54,7 @@ const COMPONENTS = [
   "OnyxTag",
   "OnyxTextarea",
   "OnyxTimer",
+  "OnyxToast",
   "OnyxTooltip",
 ] as const;
 
@@ -66,7 +69,7 @@ const filteredConfigOptions = computed(() =>
     ? configOptions.filter(({ label }) => normalizedIncludes(label, searchTerm.value))
     : configOptions,
 );
-const componentsToShow = ref([...configOptions]);
+const componentsToShow = useStorage("components-to-show", [...configOptions]);
 const componentNamesToShow = computed(() => componentsToShow.value.map((option) => option.value));
 const show = computed(() => {
   return (componentName: (typeof COMPONENTS)[number]) =>
@@ -127,6 +130,15 @@ const groupedSelectOptions: SelectOption[] = [
 
 const timerEndDate = new Date();
 timerEndDate.setHours(timerEndDate.getHours() + 2);
+
+const toast = useToast();
+
+const tableColumns = ["Fruit", "Price (€/kg)", "Inventory (kg)"];
+const tableData = [
+  { fruit: "Strawberry", price: "4.50", inventory: 200 },
+  { fruit: "Apple", price: "1.99", inventory: 3000 },
+  { fruit: "Banana", price: "3.75", inventory: 18000 },
+];
 </script>
 
 <template>
@@ -259,32 +271,33 @@ timerEndDate.setHours(timerEndDate.getHours() + 2);
           :skeleton="useSkeleton"
         />
 
-        <OnyxTable v-if="show('OnyxTable')">
-          <thead>
-            <tr>
-              <th>Fruit</th>
-              <th>Price (€/kg)</th>
-              <th>Inventory (kg)</th>
+        <template v-if="show('OnyxTable')">
+          <OnyxTable style="max-height: 250px" with-page-scrolling>
+            <template #head>
+              <tr>
+                <th v-for="col in tableColumns" :key="col">{{ col }}</th>
+              </tr>
+            </template>
+            <tr v-for="{ fruit, price, inventory } in tableData" :key="fruit">
+              <td>{{ fruit }}</td>
+              <td>{{ price }}</td>
+              <td>{{ inventory }}</td>
             </tr>
-          </thead>
-          <tbody>
-            <tr>
-              <td>Strawberry</td>
-              <td>4.50</td>
-              <td>200</td>
-            </tr>
-            <tr>
-              <td>Apple</td>
-              <td>1.99</td>
-              <td>3000</td>
-            </tr>
-            <tr>
-              <td>Banana</td>
-              <td>3.75</td>
-              <td>18000</td>
-            </tr>
-          </tbody>
-        </OnyxTable>
+          </OnyxTable>
+
+          <OnyxTable>
+            <template #head>
+              <tr>
+                <th>Empty</th>
+                <th>Table</th>
+                <th>Example</th>
+              </tr>
+            </template>
+
+            <!-- this demonstrates that the empty state works even when v-for is used -->
+            <tr v-for="(_row, index) in []" :key="index"></tr>
+          </OnyxTable>
+        </template>
 
         <OnyxTag v-if="show('OnyxTag')" label="Example tag" :icon="emojiHappy2" />
 
@@ -296,6 +309,12 @@ timerEndDate.setHours(timerEndDate.getHours() + 2);
         />
 
         <OnyxTimer v-if="show('OnyxTimer')" label="Timer" :end-time="timerEndDate" />
+
+        <OnyxButton
+          v-if="show('OnyxToast')"
+          label="Show toast"
+          @click="toast.show({ headline: 'Example toast', color: 'success' })"
+        />
 
         <OnyxTooltip v-if="show('OnyxTooltip')" text="Example tooltip text">
           Hover me to show tooltip