+++ packages/nuxt/src/runtime/locales/de-DE.js
@@ -1,0 +1,7 @@
+// @ts-check
+import { defineI18nLocale } from "#i18n";
+import deDE from "sit-onyx/locales/de-DE.json";
+
+export default defineI18nLocale(() => {
+  return { onyx: deDE };
+});