+++ packages/nuxt/src/runtime/locales/de-DE.ts
@@ -1,6 +1,0 @@
-import { defineI18nLocale } from "#i18n";
-import deDE from "sit-onyx/locales/de-DE.json";
-
-export default defineI18nLocale(() => {
-  return { onyx: deDE };
-});