+++ packages/nuxt/src/runtime/locales/en-US.js
@@ -1,0 +1,7 @@
+// @ts-check
+import { defineI18nLocale } from "#i18n";
+import enUS from "sit-onyx/locales/en-US.json";
+
+export default defineI18nLocale(() => {
+  return { onyx: enUS };
+});