+++ packages/nuxt/src/runtime/locales/en-US.ts
@@ -1,6 +1,0 @@
-import { defineI18nLocale } from "#i18n";
-import enUS from "sit-onyx/locales/en-US.json";
-
-export default defineI18nLocale(() => {
-  return { onyx: enUS };
-});