+++ packages/nuxt/src/runtime/locales/ko-KR.js
@@ -1,0 +1,7 @@
+// @ts-check
+import { defineI18nLocale } from "#i18n";
+import koKR from "sit-onyx/locales/ko-KR.json";
+
+export default defineI18nLocale(() => {
+  return { onyx: koKR };
+});