+++ packages/nuxt/src/runtime/locales/ko-KR.ts
@@ -1,6 +1,0 @@
-import { defineI18nLocale } from "#i18n";
-import koKR from "sit-onyx/locales/ko-KR.json";
-
-export default defineI18nLocale(() => {
-  return { onyx: koKR };
-});