+++ packages/sit-onyx/src/components/OnyxNavBar/modules/OnyxNavItem/future/types.ts
@@ -1,0 +1,10 @@
+export type OnyxNavItemProps = {
+  /**
+   * URL that the nav item points to.
+   */
+  href?: string;
+  /**
+   * Whether the nav item is (visually) active.
+   */
+  active?: boolean;
+};