+++ packages/sit-onyx/src/components/OnyxNavBar/modules/index.ts
@@ -1,0 +1,22 @@
+export { default as OnyxColorSchemeDialog } from "./OnyxColorSchemeDialog/OnyxColorSchemeDialog.vue";
+export * from "./OnyxColorSchemeDialog/types";
+
+export { default as OnyxColorSchemeMenuItem } from "./OnyxColorSchemeMenuItem/OnyxColorSchemeMenuItem.vue";
+export * from "./OnyxColorSchemeMenuItem/types";
+
+export { default as OnyxFlyoutMenu } from "./OnyxFlyoutMenu/OnyxFlyoutMenu.vue";
+
+export { default as OnyxNavButton } from "./OnyxNavButton/OnyxNavButton.vue";
+export * from "./OnyxNavButton/types";
+
+export { default as OnyxNavItem } from "./OnyxNavItem/OnyxNavItem.vue";
+export * from "./OnyxNavItem/types";
+
+export { default as OnyxNavSeparator } from "./OnyxNavSeparator/OnyxNavSeparator.vue";
+export * from "./OnyxNavSeparator/types";
+
+export { default as OnyxTimer } from "./OnyxTimer/OnyxTimer.vue";
+export * from "./OnyxTimer/types";
+
+export { default as OnyxUserMenu } from "./OnyxUserMenu/OnyxUserMenu.vue";
+export * from "./OnyxUserMenu/types";