+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,93 @@
 # @sit-onyx/nuxt
 
+## 1.0.0-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [4ddd145]
+  - sit-onyx@1.0.0-alpha.163
+
+## 1.0.0-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [90f9f86]
+- Updated dependencies [90f9f86]
+  - sit-onyx@1.0.0-alpha.162
+
+## 1.0.0-alpha.43
+
+### Patch Changes
+
+- Updated dependencies [4492231]
+  - sit-onyx@1.0.0-alpha.161
+
+## 1.0.0-alpha.42
+
+### Patch Changes
+
+- Updated dependencies [cd4a885]
+  - sit-onyx@1.0.0-alpha.160
+
+## 1.0.0-alpha.41
+
+### Major Changes
+
+- 981495f: Require mapping of onyx locales to the project ones.
+
+  This change was necessary because registering all languages supported by onyx would force the project to also support them due to the way locales are merged by nuxt-i18n.
+
+### Patch Changes
+
+- 981495f: Only show registration order warning if nuxt-i18n is used
+
+## 1.0.0-alpha.40
+
+### Patch Changes
+
+- Updated dependencies [760bb76]
+  - sit-onyx@1.0.0-alpha.159
+
+## 1.0.0-alpha.39
+
+### Patch Changes
+
+- 5b1aab0: fix(nuxt): i18n locales not resolvable
+
+  Prevents error on application start:
+
+  ERROR ENOENT: no such file or directory, open '/node_modules/@sit-onyx/nuxt/dist/runtime/locales/de-DE.ts'
+
+## 1.0.0-alpha.38
+
+### Patch Changes
+
+- Updated dependencies [4ee1e7f]
+  - sit-onyx@1.0.0-alpha.158
+
+## 1.0.0-alpha.37
+
+### Patch Changes
+
+- Updated dependencies [c79e491]
+- Updated dependencies [8d14b72]
+  - sit-onyx@1.0.0-alpha.157
+
+## 1.0.0-alpha.36
+
+### Patch Changes
+
+- Updated dependencies [ea9a9e7]
+  - sit-onyx@1.0.0-alpha.156
+
+## 1.0.0-alpha.35
+
+### Patch Changes
+
+- Updated dependencies [b8db0cc]
+- Updated dependencies [b8db0cc]
+  - sit-onyx@1.0.0-alpha.155
+
 ## 1.0.0-alpha.34
 
 ### Minor Changes